+++ lib/vpn/vpn_tab.dart
@@ -92,8 +92,7 @@ class VPNTab extends StatelessWidget {
                               const CDivider(height: 32.0),
                               SplitTunnelingWidget(),
                             },
-                            if (!proUser)
-                              const VPNBandwidth(),
+                            if (!proUser) const VPNBandwidth(),
                           ],
                         ),
                       ),
